requirements.txt
python-dotenv
aiohttp
pyrogram
TgCrypto
pytz
yt-dlp
requests
ffmpeg-python
instaloader